+++ languages/de/tl_content.php
@@ -1,11 +1,8 @@
 <?php 
 
-$GLOBALS['TL_LANG']['tl_content']['newslinklist']['0'] = "Nachrichtenliste";
-$GLOBALS['TL_LANG']['tl_content']['newslinklist']['1'] = "Bitte wählen Sie die anzuzeigenden Nachrichten aus.";
-$GLOBALS['TL_LANG']['tl_content']['newslinklist_tpl']['0'] = "Template";
-$GLOBALS['TL_LANG']['tl_content']['newslinklist_tpl']['1'] = "Bitte wählen Sie das Frontend-Template aus.";
-$GLOBALS['TL_LANG']['tl_content']['newslinklist_stopdate']['0'] = "bis Nachrichtendatum";
-$GLOBALS['TL_LANG']['tl_content']['newslinklist_stopdate']['1'] = "Nur Nachrichten bis zu diesem Datum anzeigen.";
+$GLOBALS['TL_LANG']['tl_content']['newslinklist'] = array('Nachrichtenliste', 'Bitte wählen Sie die anzuzeigenden Nachrichten aus.';
+$GLOBALS['TL_LANG']['tl_content']['newslinklist_tpl'] = array('Template', 'Bitte wählen Sie das Frontend-Template aus.';
+$GLOBALS['TL_LANG']['tl_content']['newslinklist_stopdate'] = array('bis Nachrichtendatum', 'Nur Nachrichten bis zu diesem Datum anzeigen.');
 
 $GLOBALS['TL_LANG']['tl_content']['nll_legend'] = 'Einstellungen der Nachrichten-Linkliste';